requirements.txt
py-cord==2.6.0
pynacl==1.5.0
numpy==1.26.3
scipy==1.11.4
pydub==0.25.1
fastapi==0.109.2
uvicorn==0.27.0
python-multipart==0.0.6
pyjwt==2.8.0
python-dotenv==1.0.0
aiosqlite==0.19.0